requirements.txt
llama-cpp-python
semantic-router[local]

